==> 3. textual-processing.py <==
"""

TEXTUAL PROCESSING FILE (NLP)

ALEXANDER LUND - THESIS ESCP

"""
import pandas as pd
import nltk
from nltk.corpus import stopwords
import spacy
from sklearn.feature_extraction.text import TfidfVectorizer
import pysentiment2 as ps
import demoji
from collections import OrderedDict
from collections import defaultdict
from collections import Counter
from textblob import TextBlob

############# REDDIT ####################

df = pd.read_csv("2. reddit-sentiments.csv")

stopw = stopwords.words("english")

df["body"] = df["body"].astype(str)

# drop duplicates
df = df.drop(df.index[[3009]])

# clean up of issue derived from scraping (certain rows have timestamp and body reversed)
df['timestamp_2'] = df["timestamp"]
df['body_2'] = df["body"]

df.loc[df['body'].str.startswith(("12","13","14","15","16")),'timestamp_2'] = df['body']
df.loc[df['body'].str.startswith(("12","13","14","15","16")),'body_2'] = df['timestamp']
df.drop(["body", "timestamp"], axis=1, inplace =True)

df.rename(columns={"body_2":"body", "timestamp_2":"timestamp"}, inplace=True)

df.drop(df.loc[df['body']=="1616581624.0"].index, inplace=True)
df.drop(df.loc[df['body']=="1616962909.0"].index, inplace=True)


# remove stopwords, punctuation and further cleaning
df["cleaned body"] = df["body"].apply(lambda x: " ".join([word for word in x.split() if word.lower() not in (stopw)]))

df["cleaned body"] = df["cleaned body"].str.replace('[^\w\s]','')

df["cleaned body"] = df["cleaned body"].apply(lambda x: " ".join(x.lower() for x in x.split()))

df = df[df["cleaned body"].str.contains("removed|deleted")==False]

# lemmatization
nlp = spacy.load("en_core_web_sm")

def space(comment):
    doc = nlp(comment)
    return " ".join([token.lemma_ for token in doc])

df["cleaned body"] = df["cleaned body"].apply(space)


# define personal ESG dictionaries, load LoughranMcDonald dictionary
positive_sustainable_dictionary = ["environmental", "social", "governance", "planet", "positive impact", "index", "sustainability", "green", "sustainable investing", "responsible investment", "impact investing", "sustainable finance", "green finance", "corporate social responsibility", "green" "esg investing", "socially responsible investing", "climate funds", "portfolio carbon footprint", "carbon finance", "positive screening", "best in class ESG", "sustainability index", "thematic investing", "triple bottom line", "solar", "wind", "carbon free", "paris agreement", "renewable", "alternative", "green portfolio", "natural", "ocean", "reform", "reforestation", "environmentalist", "microfinance"]

negative_sustainable_dictionary = ["negative impact", "greenwashing", "increase carbon emissions", "increase co2 emissions", "crisis", "pollution", "greenwash", "burn" "burning fossile fuel", "cement", "deforestation", "harvest", "land abuse", "destruction", "lack", "inaction", "unawareness", "not care", "abusive", "consumption",  "negation", "sin",  "vice"]

positive_louhgranmcdonald_dictionary = ["able", "abundance", "abundant", "acclaimed", "accomplish",	"accomplished",	"accomplishes",	"accomplishing", "accomplishment", "accomplishments", "achieve",	"achieved",	"achievement",	"achievements",	"achieves",	"achieving",	"adequately",	"advancement",	"advancements",	"advances",	"advancing",	"advantage",	"advantaged",	"advantageous",	"advantageously",	"advantages",	"alliance",	"alliances",	"assure",	"assured",	"assures",	"assuring",	"attain",	"attained",	"attaining",	"attainment",	"attainments",	"attains",	"attractive",	"attractiveness",	"beautiful",	"beautifully",	"beneficial",	"beneficially",	"benefit",	"benefited",	"benefiting",	"benefitted",	"benefitting",	"best",	"better",	"bolstered",	"bolstering",	"bolsters",	"boom",	"booming",	"boost",	"boosted",	"breakthrough",	"breakthroughs",	"brilliant",	"charitable",	"collaborate",	"collaborated",	"collaborates",	"collaborating",	"collaboration",	"collaborations",	"collaborative",	"collaborator",	"collaborators",	"compliment",	"complimentary",	"complimented",	"complimenting",	"compliments",	"conclusive",	"conclusively",	"conducive",	"confident",	"constructive",	"constructively",	"courteous",	"creative",	"creatively",	"creativeness",	"creativity",	"delight",	"delighted",	"delightful",	"delightfully",	"delighting",	"delights",	"dependability",	"dependable",	"desirable",	"desired",	"despite",	"destined",	"diligent",	"diligently",	"distinction",	"distinctions",	"distinctive",	"distinctively",	"distinctiveness",	"dream",	"easier",	"easily",	"easy",	"effective",	"efficiencies",	"efficiency",	"efficient",	"efficiently",	"empower",	"empowered",	"empowering",	"empowers",	"enable",	"enabled",	"enables",	"enabling",	"encouraged",	"encouragement",	"encourages",	"encouraging",	"enhance",	"enhanced",	"enhancement",	"enhancements",	"enhances",	"enhancing",	"enjoy",	"enjoyable",	"enjoyably",	"enjoyed",	"enjoying",	"enjoyment",	"enjoys",	"enthusiasm",	"enthusiastic",	"enthusiastically",	"excellence",	"excellent",	"excelling",	"excels",	"exceptional",	"exceptionally",	"excited",	"excitement",	"exciting",	"exclusive",	"exclusively",	"exclusiveness",	"exclusives",	"exclusivity",	"exemplary",	"fantastic",	"favorable",	"favorably",	"favored",	"favoring",	"favorite",	"favorites",	"friendly",	"gain",	"gained",	"gaining",	"gains",	"good",	"great",	"greater",	"greatest",	"greatly",	"greatness",	"happiest",	"happily",	"happiness",	"happy",	"highest",	"honor",	"honorable",	"honored",	"honoring",	"honors",	"ideal",	"impress",	"impressed",	"impresses",	"impressing",	"impressive",	"impressively",	"improve",	"improved",	"improvement",	"improvements",	"improves",	"improving",	"incredible",	"incredibly",	"influential",	"informative",	"ingenuity",	"innovate",	"innovated",	"innovates",	"innovating",	"innovation",	"innovations",	"innovative",	"innovativeness",	"innovator",	"innovators",	"insightful",	"inspiration",	"inspirational",	"integrity",	"invent",	"invented",	"inventing",	"invention",	"inventions",	"inventive",	"inventiveness",	"inventor",	"inventors",	"leadership",	"leading",	"loyal",	"lucrative",	"meritorious",	"opportunities",	"opportunity",	"optimistic",	"outperform",	"outperformed",	"outperforming",	"outperforms",	"perfect",	"perfected",	"perfectly",	"perfects",	"pleasant",	"pleasantly",	"pleased",	"pleasure",	"plentiful",	"popular",	"popularity",	"positive",	"positively",	"preeminence",	"preeminent",	"premier",	"premiere",	"prestige",	"prestigious",	"proactive",	"proactively",	"proficiency",	"proficient",	"proficiently",	"profitability",	"profitable",	"profitably",	"progress",	"progressed",	"progresses",	"progressing",	"prospered",	"prospering",	"prosperity",	"prosperous",	"prospers",	"rebound",	"rebounded",	"rebounding",	"receptive",	"regain",	"regained",	"regaining",	"resolve",	"revolutionize",	"revolutionized",	"revolutionizes",	"revolutionizing",	"reward",	"rewarded",	"rewarding",	"rewards",	"satisfaction",	"satisfactorily",	"satisfactory",	"satisfied",	"satisfies",	"satisfy",	"satisfying",	"smooth",	"smoothing",	"smoothly",	"smooths",	"solves",	"solving",	"spectacular",	"spectacularly",	"stability",	"stabilization",	"stabilizations",	"stabilize",	"stabilized",	"stabilizes",	"stabilizing",	"stable",	"strength",	"strengthen",	"strengthened",	"strengthening",	"strengthens",	"strengths",	"strong",	"stronger",	"strongest",	"succeed",	"succeeded",	"succeeding",	"succeeds",	"success",	"successes",	"successful",	"successfully",	"superior",	"surpass",	"surpassed",	"surpasses",	"surpassing",	"transparency",	"tremendous",	"tremendously",	"unmatched",	"unparalleled",	"unsurpassed",	"upturn",	"upturns",	"valuable",	"versatile",	"versatility",	"vibrancy",	"vibrant",	"win",	"winner",	"winners",	"winning",	"worthy"]

negative_louhgranmcdonald_dictionary = ["abandon",	"abandoned",	"abandoning",	"abandonment",	"abandonments",	"abandons",	"abdicated",	"abdicates",	"abdicating",	"abdication",	"abdications",	"aberrant",	"aberration",	"aberrational",	"aberrations",	"abetting",	"abnormal",	"abnormalities",	"abnormality",	"abnormally",	"abolish",	"abolished",	"abolishes",	"abolishing",	"abrogate",	"abrogated",	"abrogates",	"abrogating",	"abrogation",	"abrogations",	"abrupt",	"abruptly",	"abruptness",	"absence",	"absences",	"absenteeism",	"abuse",	"abused",	"abuses",	"abusing",	"abusive",	"abusively",	"abusiveness",	"accident",	"accidental",	"accidentally",	"accidents",	"accusation",	"accusations",	"accuse",	"accused",	"accuses",	"accusing",	"acquiesce",	"acquiesced",	"acquiesces",	"acquiescing",	"acquit",	"acquits",	"acquittal",	"acquittals",	"acquitted",	"acquitting",	"adulterate",	"adulterated",	"adulterating",	"adulteration",	"adulterations",	"adversarial",	"adversaries",	"adversary",	"adverse",	"adversely",	"adversities",	"adversity",	"aftermath",	"aftermaths",	"against",	"aggravate",	"aggravated",	"aggravates",	"aggravating",	"aggravation",	"aggravations",	"alerted",	"alerting",	"alienate",	"alienated",	"alienates",	"alienating",	"alienation",	"alienations",	"allegation",	"allegations",	"allege",	"alleged",	"allegedly",	"alleges",	"alleging",	"annoy",	"annoyance",	"annoyances",	"annoyed",	"annoying",	"annoys",	"annul",	"annulled",	"annulling",	"annulment",	"annulments",	"annuls",	"anomalies",	"anomalous",	"anomalously",	"anomaly",	"anticompetitive",	"antitrust",	"argue",	"argued",	"arguing",	"argument",	"argumentative",	"arguments",	"arrearage",	"arrearages",	"arrears",	"arrest",	"arrested",	"arrests",	"artificially",	"assault",	"assaulted",	"assaulting",	"assaults",	"assertions",	"attrition",	"aversely",	"backdating",	"bad",	"bail",	"bailout",	"balk",	"balked",	"bankrupt",	"bankruptcies",	"bankruptcy",	"bankrupted",	"bankrupting",	"bankrupts",	"bans",	"barred",	"barrier",	"barriers",	"bottleneck",	"bottlenecks",	"boycott",	"boycotted",	"boycotting",	"boycotts",	"breach",	"breached",	"breaches",	"breaching",	"break",	"breakage",	"breakages",	"breakdown",	"breakdowns",	"breaking",	"breaks",	"bribe",	"bribed",	"briberies",	"bribery",	"bribes",	"bribing",	"bridge",	"broken",	"burden",	"burdened",	"burdening",	"burdens",	"burdensome",	"burned",	"calamities",	"calamitous",	"calamity",	"cancel",	"canceled",	"canceling",	"cancellation",	"cancellations",	"cancelled",	"cancelling",	"cancels",	"careless",	"carelessly",	"carelessness",	"catastrophe",	"catastrophes",	"catastrophic",	"catastrophically",	"caution",	"cautionary",	"cautioned",	"cautioning",	"cautions",	"cease",	"ceased",	"ceases",	"ceasing",	"censure",	"censured",	"censures",	"censuring",	"challenge",	"challenged",	"challenges",	"challenging",	"chargeoffs",	"circumvent",	"circumvented",	"circumventing",	"circumvention",	"circumventions",	"circumvents",	"claiming",	"claims",	"clawback",	"closed",	"closeout",	"closeouts",	"closing",	"closings",	"closure",	"closures",	"coerce",	"coerced",	"coerces",	"coercing",	"coercion",	"coercive",	"collapse",	"collapsed",	"collapses",	"collapsing",	"collision",	"collisions",	"collude",	"colluded",	"colludes",	"colluding",	"collusion",	"collusions",	"collusive",	"complain",	"complained",	"complaining",	"complains",	"complaint",	"complaints",	"complicate",	"complicated",	"complicates",	"complicating",	"complication",	"complications",	"compulsion",	"concealed",	"concealing",	"concede",	"conceded",	"concedes",	"conceding",	"concern",	"concerned",	"concerns",	"conciliating",	"conciliation",	"conciliations",	"condemn",	"condemnation",	"condemnations",	"condemned",	"condemning",	"condemns",	"condone",	"condoned",	"confess",	"confessed",	"confesses",	"confessing",	"confession",	"confine",	"confined",	"confinement",	"confinements",	"confines",	"confining",	"confiscate",	"confiscated",	"confiscates",	"confiscating",	"confiscation",	"confiscations",	"conflict",	"conflicted",	"conflicting",	"conflicts",	"confront",	"confrontation",	"confrontational",	"confrontations",	"confronted",	"confronting",	"confronts",	"confuse",	"confused",	"confuses",	"confusing",	"confusingly",	"confusion",	"conspiracies",	"conspiracy",	"conspirator",	"conspiratorial",	"conspirators",	"conspire",	"conspired",	"conspires",	"conspiring",	"contempt",	"contend",	"contended",	"contending",	"contends",	"contention",	"contentions",	"contentious",	"contentiously",	"contested",	"contesting",	"contraction",	"contractions",	"contradict",	"contradicted",	"contradicting",	"contradiction",	"contradictions",	"contradictory",	"contradicts",	"contrary",	"controversial",	"controversies",	"controversy",	"convict",	"convicted",	"convicting",	"conviction",	"convictions",	"corrected",	"correcting",	"correction",	"corrections",	"corrects",	"corrupt",	"corrupted",	"corrupting",	"corruption",	"corruptions",	"corruptly",	"corruptness",	"costly",	"counterclaim",	"counterclaimed",	"counterclaiming",	"counterclaims",	"counterfeit",	"counterfeited",	"counterfeiter",	"counterfeiters",	"counterfeiting",	"counterfeits",	"countermeasure",	"countermeasures",	"crime",	"crimes",	"criminal",	"criminally",	"criminals",	"crises",	"crisis",	"critical",	"critically",	"criticism",	"criticisms",	"criticize",	"criticized",	"criticizes",	"criticizing",	"crucial",	"crucially",	"culpability",	"culpable",	"culpably",	"cumbersome",	"curtail",	"curtailed",	"curtailing",	"curtailment",	"curtailments",	"curtails",	"cut",	"cutback",	"cutbacks",	"cyberattack",	"cyberattacks",	"cyberbullying",	"cybercrime",	"cybercrimes",	"cybercriminal",	"cybercriminals",	"damage",	"damaged",	"damages",	"damaging",	"dampen",	"dampened",	"danger",	"dangerous",	"dangerously",	"dangers",	"deadlock",	"deadlocked",	"deadlocking",	"deadlocks",	"deadweight",	"deadweights",	"debarment",	"debarments",	"debarred",	"deceased",	"deceit",	"deceitful",	"deceitfulness",	"deceive",	"deceived",	"deceives",	"deceiving",	"deception",	"deceptions",	"deceptive",	"deceptively",	"decline",	"declined",	"declines",	"declining",	"deface",	"defaced",	"defacement",	"defamation",	"defamations",	"defamatory",	"defame",	"defamed",	"defames",	"defaming",	"default",	"defaulted",	"defaulting",	"defaults",	"defeat",	"defeated",	"defeating",	"defeats",	"defect",	"defective",	"defects",	"defend",	"defendant",	"defendants",	"defended",	"defending",	"defends",	"defensive",	"defer",	"deficiencies",	"deficiency",	"deficient",	"deficit",	"deficits",	"defraud",	"defrauded",	"defrauding",	"defrauds",	"defunct",	"degradation",	"degradations",	"degrade",	"degraded",	"degrades",	"degrading",	"delay",	"delayed",	"delaying",	"delays",	"deleterious",	"deliberate",	"deliberated",	"deliberately",	"delinquencies",	"delinquency",	"delinquent",	"delinquently",	"delinquents",	"delist",	"delisted",	"delisting",	"delists",	"demise",	"demised",	"demises",	"demising",	"demolish",	"demolished",	"demolishes",	"demolishing",	"demolition",	"demolitions",	"demote",	"demoted",	"demotes",	"demoting",	"demotion",	"demotions",	"denial",	"denials",	"denied",	"denies",	"denigrate",	"denigrated",	"denigrates",	"denigrating",	"denigration",	"deny",	"denying",	"deplete",	"depleted",	"depletes",	"depleting",	"depletion",	"depletions",	"deprecation",	"depress",	"depressed",	"depresses",	"depressing",	"deprivation",	"deprive",	"deprived",	"deprives",	"depriving",	"derelict",	"dereliction",	"derogatory",	"destabilization",	"destabilize",	"destabilized",	"destabilizing",	"destroy",	"destroyed",	"destroying",	"destroys",	"destruction",	"destructive",	"detain",	"detained",	"detention",	"detentions",	"deter",	"deteriorate",	"deteriorated",	"deteriorates",	"deteriorating",	"deterioration",	"deteriorations",	"deterred",	"deterrence",	"deterrences",	"deterrent",	"deterrents",	"deterring",	"deters",	"detract",	"detracted",	"detracting",	"detriment",	"detrimental",	"detrimentally",	"detriments",	"devalue",	"devalued",	"devalues",	"devaluing",	"devastate",	"devastated",	"devastating",	"devastation",	"deviate",	"deviated",	"deviates",	"deviating",	"deviation",	"deviations",	"devolve",	"devolved",	"devolves",	"devolving",	"difficult",	"difficulties",	"difficultly",	"difficulty",	"diminish",	"diminished",	"diminishes",	"diminishing",	"diminution",	"disadvantage",	"disadvantaged",	"disadvantageous",	"disadvantages",	"disaffiliation",	"disagree",	"disagreeable",	"disagreed",	"disagreeing",	"disagreement",	"disagreements",	"disagrees",	"disallow",	"disallowance",	"disallowances",	"disallowed",	"disallowing",	"disallows",	"disappear",	"disappearance",	"disappearances",	"disappeared",	"disappearing",	"disappears",	"disappoint",	"disappointed",	"disappointing",	"disappointingly",	"disappointment",	"disappointments",	"disappoints",	"disapproval",	"disapprovals",	"disapprove",	"disapproved",	"disapproves",	"disapproving",	"disassociates",	"disassociating",	"disassociation",	"disassociations",	"disaster",	"disasters",	"disastrous",	"disastrously",	"disavow",	"disavowal",	"disavowed",	"disavowing",	"disavows",	"disciplinary",	"disclaim",	"disclaimed",	"disclaimer",	"disclaimers",	"disclaiming",	"disclaims",	"disclose",	"disclosed",	"discloses",	"disclosing",	"discontinuance",	"discontinuances",	"discontinuation",	"discontinuations",	"discontinue",	"discontinued",	"discontinues",	"discontinuing",	"discourage",	"discouraged",	"discourages",	"discouraging",	"discredit",	"discredited",	"discrediting",	"discredits",	"discrepancies",	"discrepancy",	"disfavor",	"disfavored",	"disfavoring",	"disfavors",	"disgorge",	"disgorged",	"disgorgement",	"disgorgements",	"disgorges",	"disgorging",	"disgrace",	"disgraceful",	"disgracefully",	"dishonest",	"dishonestly",	"dishonesty",	"dishonor",	"dishonorable",	"dishonorably",	"dishonored",	"dishonoring",	"dishonors",	"disincentives",	"disinterested",	"disinterestedly",	"disinterestedness",	"disloyal",	"disloyally",	"disloyalty",	"dismal",	"dismally",	"dismiss",	"dismissal",	"dismissals",	"dismissed",	"dismisses",	"dismissing",	"disorderly",	"disparage",	"disparaged",	"disparagement",	"disparagements",	"disparages",	"disparaging",	"disparagingly",	"disparities",	"disparity",	"displace",	"displaced",	"displacement",	"displacements",	"displaces",	"displacing",	"dispose",	"dispossess",	"dispossessed",	"dispossesses",	"dispossessing",	"disproportion",	"disproportional",	"disproportionate",	"disproportionately",	"dispute",	"disputed",	"disputes",	"disputing",	"disqualification",	"disqualifications",	"disqualified",	"disqualifies",	"disqualify",	"disqualifying",	"disregard",	"disregarded",	"disregarding",	"disregards",	"disreputable",	"disrepute",	"disrupt",	"disrupted",	"disrupting",	"disruption",	"disruptions",	"disruptive",	"disrupts",	"dissatisfaction",	"dissatisfied",	"dissent",	"dissented",	"dissenter",	"dissenters",	"dissenting",	"dissents",	"dissident",	"dissidents",	"dissolution",	"dissolutions",	"distort",	"distorted",	"distorting",	"distortion",	"distortions",	"distorts",	"distract",	"distracted",	"distracting",	"distraction",	"distractions",	"distracts",	"distress",	"distressed",	"disturb",	"disturbance",	"disturbances",	"disturbed",	"disturbing",	"disturbs",	"diversion",	"divert",	"diverted",	"diverting",	"diverts",	"divest",	"divested",	"divesting",	"divestiture",	"divestitures",	"divestment",	"divestments",	"divests",	"divorce",	"divorced",	"divulge",	"divulged",	"divulges",	"divulging",	"doubt",	"doubted",	"doubtful",	"doubts",	"downgrade",	"downgraded",	"downgrades",	"downgrading",	"downsize",	"downsized",	"downsizes",	"downsizing",	"downsizings",	"downtime",	"downtimes",	"downturn",	"downturns",	"downward",	"downwards",	"drag",	"drastic",	"drastically",	"drawback",	"drawbacks",	"dropped",	"drought",	"droughts",	"duress",	"dysfunction",	"dysfunctional",	"dysfunctions",	"easing",	"egregious",	"egregiously",	"embargo",	"embargoed",	"embargoes",	"embargoing",	"embarrass",	"embarrassed",	"embarrasses",	"embarrassing",	"embarrassment",	"embarrassments",	"embezzle",	"embezzled",	"embezzlement",	"embezzlements",	"embezzler",	"embezzles",	"embezzling",	"encroach",	"encroached",	"encroaches",	"encroaching",	"encroachment",	"encroachments",	"encumber",	"encumbered",	"encumbering",	"encumbers",	"encumbrance",	"encumbrances",	"endanger",	"endangered",	"endangering",	"endangerment",	"endangers",	"enjoin",	"enjoined",	"enjoining",	"enjoins",	"erode",	"eroded",	"erodes",	"eroding",	"erosion",	"erratic",	"erratically",	"erred",	"erring",	"erroneous",	"erroneously",	"error",	"errors",	"errs",	"escalate",	"escalated",	"escalates",	"escalating",	"evade",	"evaded",	"evades",	"evading",	"evasion",	"evasions",	"evasive",	"evict",	"evicted",	"evicting",	"eviction",	"evictions",	"evicts",	"exacerbate",	"exacerbated",	"exacerbates",	"exacerbating",	"exacerbation",	"exacerbations",	"exaggerate",	"exaggerated",	"exaggerates",	"exaggerating",	"exaggeration",	"excessive",	"excessively",	"exculpate",	"exculpated",	"exculpates",	"exculpating",	"exculpation",	"exculpations",	"exculpatory",	"exonerate",	"exonerated",	"exonerates",	"exonerating",	"exoneration",	"exonerations",	"exploit",	"exploitation",	"exploitations",	"exploitative",	"exploited",	"exploiting",	"exploits",	"expose",	"exposed",	"exposes",	"exposing",	"expropriate",	"expropriated",	"expropriates",	"expropriating",	"expropriation",	"expropriations",	"expulsion",	"expulsions",	"extenuating",	"fail",	"failed",	"failing",	"failings",	"fails",	"failure",	"failures",	"fallout",	"false",	"falsely",	"falsification",	"falsifications",	"falsified",	"falsifies",	"falsify",	"falsifying",	"falsity",	"fatalities",	"fatality",	"fatally",	"fault",	"faulted",	"faults",	"faulty",	"fear",	"fears",	"felonies",	"felonious",	"felony",	"fictitious",	"fined",	"fines",	"fired",	"firing",	"flaw",	"flawed",	"flaws",	"forbid",	"forbidden",	"forbidding",	"forbids",	"force",	"forced",	"forcing",	"foreclose",	"foreclosed",	"forecloses",	"foreclosing",	"foreclosure",	"foreclosures",	"forego",	"foregoes",	"foregone",	"forestall",	"forestalled",	"forestalling",	"forestalls",	"forfeit",	"forfeited",	"forfeiting",	"forfeits",	"forfeiture",	"forfeitures",	"forgers",	"forgery",	"fraud",	"frauds",	"fraudulence",	"fraudulent",	"fraudulently",	"frivolous",	"frivolously",	"frustrate",	"frustrated",	"frustrates",	"frustrating",	"frustratingly",	"frustration",	"frustrations",	"fugitive",	"fugitives",	"gratuitous",	"gratuitously",	"grievance",	"grievances",	"grossly",	"groundless",	"guilty",	"halt",	"halted",	"hamper",	"hampered",	"hampering",	"hampers",	"harass",	"harassed",	"harassing",	"harassment",	"hardship",	"hardships",	"harm",	"harmed",	"harmful",	"harmfully",	"harming",	"harms",	"harsh",	"harsher",	"harshest",	"harshly",	"harshness",	"hazard",	"hazardous",	"hazards",	"hinder",	"hindered",	"hindering",	"hinders",	"hindrance",	"hindrances",	"hostile",	"hostility",	"hurt",	"hurting",	"idle",	"idled",	"idling",	"ignore",	"ignored",	"ignores",	"ignoring",	"ill",	"illegal",	"illegalities",	"illegality",	"illegally",	"illegible",	"illicit",	"illicitly",	"illiquid",	"illiquidity",	"imbalance",	"imbalances",	"immature",	"immoral",	"impair",	"impaired",	"impairing",	"impairment",	"impairments",	"impairs",	"impasse",	"impasses",	"impede",	"impeded",	"impedes",	"impediment",	"impediments",	"impeding",	"impending",	"imperative",	"imperfection",	"imperfections",	"imperil",	"impermissible",	"implicate",	"implicated",	"implicates",	"implicating",	"impossibility",	"impossible",	"impound",	"impounded",	"impounding",	"impounds",	"impracticable",	"impractical",	"impracticalities",	"impracticality",	"imprisonment",	"improper",	"improperly",	"improprieties",	"impropriety",	"imprudent",	"imprudently",	"inability",	"inaccessible",	"inaccuracies",	"inaccuracy",	"inaccurate",	"inaccurately",	"inaction",	"inactions",	"inactivate",	"inactivated",	"inactivates",	"inactivating",	"inactivation",	"inactivations",	"inactivity",	"inadequacies",	"inadequacy",	"inadequate",	"inadequately",	"inadvertent",	"inadvertently",	"inadvisability",	"inadvisable",	"inappropriate",	"inappropriately",	"inattention",	"incapable",	"incapacitated",	"incapacity",	"incarcerate",	"incarcerated",	"incarcerates",	"incarcerating",	"incarceration",	"incarcerations",	"incidence",	"incidences",	"incident",	"incidents",	"incompatibilities",	"incompatibility",	"incompatible",	"incompetence",	"incompetency",	"incompetent",	"incompetently",	"incompetents",	"incomplete",	"incompletely",	"incompleteness",	"inconclusive",	"inconsistencies",	"inconsistency",	"inconsistent",	"inconsistently",	"inconvenience",	"inconveniences",	"inconvenient",	"incorrect",	"incorrectly",	"incorrectness",	"indecency",	"indecent",	"indefeasible",	"indefeasibly",	"indict",	"indictable",	"indicted",	"indicting",	"indictment",	"indictments",	"ineffective",	"ineffectively",	"ineffectiveness",	"inefficiencies",	"inefficiency",	"inefficient",	"inefficiently",	"ineligibility",	"ineligible",	"inequitable",	"inequitably",	"inequities",	"inequity",	"inevitable",	"inexperience",	"inexperienced",	"inferior",	"inflicted",	"infraction",	"infractions",	"infringe",	"infringed",	"infringement",	"infringements",	"infringes",	"infringing",	"inhibited",	"inimical",	"injunction",	"injunctions",	"injure",	"injured",	"injures",	"injuries",	"injuring",	"injurious",	"injury",	"inordinate",	"inordinately",	"inquiry",	"insecure",	"insensitive",	"insolvencies",	"insolvency",	"insolvent",	"instability",	"insubordination",	"insufficiency",	"insufficient",	"insufficiently",	"insurrection",	"insurrections",	"intentional",	"interfere",	"interfered",	"interference",	"interferences",	"interferes",	"interfering",	"intermittent",	"intermittently",	"interrupt",	"interrupted",	"interrupting",	"interruption",	"interruptions",	"interrupts",	"intimidation",	"intrusion",	"invalid",	"invalidate",	"invalidated",	"invalidates",	"invalidating",	"invalidation",	"invalidity",	"investigate",	"investigated",	"investigates",	"investigating",	"investigation",	"investigations",	"involuntarily",	"involuntary",	"irreconcilable",	"irreconcilably",	"irrecoverable",	"irrecoverably",	"irregular",	"irregularities",	"irregularity",	"irregularly",	"irreparable",	"irreparably",	"irreversible",	"jeopardize",	"jeopardized",	"justifiable",	"kickback",	"kickbacks",	"knowingly",	"lack",	"lacked",	"lacking",	"lackluster",	"lacks",	"lag",	"lagged",	"lagging",	"lags",	"lapse",	"lapsed",	"lapses",	"lapsing",	"late",	"laundering",	"layoff",	"layoffs",	"lie",	"limitation",	"limitations",	"lingering",	"liquidate",	"liquidated",	"liquidates",	"liquidating",	"liquidation",	"liquidations",	"liquidator",	"liquidators",	"litigant",	"litigants",	"litigate",	"litigated",	"litigates",	"litigating",	"litigation",	"litigations",	"lockout",	"lockouts",	"lose",	"loses",	"losing",	"loss",	"losses",	"lost",	"lying",	"malfeasance",	"malfunction",	"malfunctioned",	"malfunctioning",	"malfunctions",	"malice",	"malicious",	"maliciously",	"malpractice",	"manipulate",	"manipulated",	"manipulates",	"manipulating",	"manipulation",	"manipulations",	"manipulative",	"markdown",	"markdowns",	"misapplication",	"misapplications",	"misapplied",	"misapplies",	"misapply",	"misapplying",	"misappropriate",	"misappropriated",	"misappropriates",	"misappropriating",	"misappropriation",	"misappropriations",	"misbranded",	"miscalculate",	"miscalculated",	"miscalculates",	"miscalculating",	"miscalculation",	"miscalculations",	"mischaracterization",	"mischief",	"misclassification",	"misclassifications",	"misclassified",	"misclassify",	"miscommunication",	"misconduct",	"misdated",	"misdemeanor",	"misdemeanors",	"misdirected",	"mishandle",	"mishandled",	"mishandles",	"mishandling",	"misinform",	"misinformation",	"misinformed",	"misinforming",	"misinforms",	"misinterpret",	"misinterpretation",	"misinterpretations",	"misinterpreted",	"misinterpreting",	"misinterprets",	"misjudge",	"misjudged",	"misjudges",	"misjudging",	"misjudgment",	"misjudgments",	"mislabel",	"mislabeled",	"mislabeling",	"mislabelled",	"mislabels",	"mislead",	"misleading",	"misleadingly",	"misleads",	"misled",	"mismanage",	"mismanaged",	"mismanagement",	"mismanages",	"mismanaging",	"mismatch",	"mismatched",	"mismatches",	"mismatching",	"misplaced",	"misprice",	"mispricing",	"mispricings",	"misrepresent",	"misrepresentation",	"misrepresentations",	"misrepresented",	"misrepresenting",	"misrepresents",	"miss",	"missed",	"misses",	"misstate",	"misstated",	"misstatement",	"misstatements",	"misstates",	"misstating",	"misstep",	"missteps",	"mistake",	"mistaken",	"mistakenly",	"mistakes",	"mistaking",	"mistrial",	"mistrials",	"misunderstand",	"misunderstanding",	"misunderstandings",	"misunderstood",	"misuse",	"misused",	"misuses",	"misusing",	"monopolistic",	"monopolists",	"monopolization",	"monopolize",	"monopolized",	"monopolizes",	"monopolizing",	"monopoly",	"moratoria",	"moratorium",	"moratoriums",	"mothballed",	"mothballing",	"negative",	"negatively",	"negatives",	"neglect",	"neglected",	"neglectful",	"neglecting",	"neglects",	"negligence",	"negligences",	"negligent",	"negligently",	"nonattainment",	"noncompetitive",	"noncompliance",	"noncompliances",	"noncompliant",	"noncomplying",	"nonconforming",	"nonconformities",	"nonconformity",	"nondisclosure",	"nonfunctional",	"nonpayment",	"nonpayments",	"nonperformance",	"nonperformances",	"nonperforming",	"nonproducing",	"nonproductive",	"nonrecoverable",	"nonrenewal",	"nuisance",	"nuisances",	"nullification",	"nullifications",	"nullified",	"nullifies",	"nullify",	"nullifying",	"objected",	"objecting",	"objection",	"objectionable",	"objectionably",	"objections",	"obscene",	"obscenity",	"obsolescence",	"obsolete",	"obstacle",	"obstacles",	"obstruct",	"obstructed",	"obstructing",	"obstruction",	"obstructions",	"offence",	"offences",	"offend",	"offended",	"offender",	"offenders",	"offending",	"offends",	"omission",	"omissions",	"omit",	"omits",	"omitted",	"omitting",	"onerous",	"opportunistic",	"opportunistically",	"oppose",	"opposed",	"opposes",	"opposing",	"opposition",	"oppositions",	"outage",	"outages",	"outdated",	"outmoded",	"overage",	"overages",	"overbuild",	"overbuilding",	"overbuilds",	"overbuilt",	"overburden",	"overburdened",	"overburdening",	"overcapacities",	"overcapacity",	"overcharge",	"overcharged",	"overcharges",	"overcharging",	"overcome",	"overcomes",	"overcoming",	"overdue",	"overestimate",	"overestimated",	"overestimates",	"overestimating",	"overestimation",	"overestimations",	"overload",	"overloaded",	"overloading",	"overloads",	"overlook",	"overlooked",	"overlooking",	"overlooks",	"overpaid",	"overpayment",	"overpayments",	"overproduced",	"overproduces",	"overproducing",	"overproduction",	"overrun",	"overrunning",	"overruns",	"overshadow",	"overshadowed",	"overshadowing",	"overshadows",	"overstate",	"overstated",	"overstatement",	"overstatements",	"overstates",	"overstating",	"oversupplied",	"oversupplies",	"oversupply",	"oversupplying",	"overtly",	"overturn",	"overturned",	"overturning",	"overturns",	"overvalue",	"overvalued",	"overvaluing",	"panic",	"panics",	"penalize",	"penalized",	"penalizes",	"penalizing",	"penalties",	"penalty",	"peril",	"perils",	"perjury",	"perpetrate",	"perpetrated",	"perpetrates",	"perpetrating",	"perpetration",	"persist",	"persisted",	"persistence",	"persistent",	"persistently",	"persisting",	"persists",	"pervasive",	"pervasively",	"pervasiveness",	"petty",	"picket",	"picketed",	"picketing",	"plaintiff",	"plaintiffs",	"plea",	"plead",	"pleaded",	"pleading",	"pleadings",	"pleads",	"pleas",	"pled",	"poor",	"poorly",	"poses",	"posing",	"postpone",	"postponed",	"postponement",	"postponements",	"postpones",	"postponing",	"precipitated",	"precipitous",	"precipitously",	"preclude",	"precluded",	"precludes",	"precluding",	"predatory",	"prejudice",	"prejudiced",	"prejudices",	"prejudicial",	"prejudicing",	"premature",	"prematurely",	"pressing",	"pretrial",	"preventing",	"prevention",	"prevents",	"problem",	"problematic",	"problematical",	"problems",	"prolong",	"prolongation",	"prolongations",	"prolonged",	"prolonging",	"prolongs",	"prone",	"prosecute",	"prosecuted",	"prosecutes",	"prosecuting",	"prosecution",	"prosecutions",	"protest",	"protested",	"protester",	"protesters",	"protesting",	"protestor",	"protestors",	"protests",	"protracted",	"protraction",	"provoke",	"provoked",	"provokes",	"provoking",	"punished",	"punishes",	"punishing",	"punishment",	"punishments",	"punitive",	"purport",	"purported",	"purportedly",	"purporting",	"purports",	"question",	"questionable",	"questionably",	"questioned",	"questioning",	"questions",	"quit",	"quitting",	"racketeer",	"racketeering",	"rationalization",	"rationalizations",	"rationalize",	"rationalized",	"rationalizes",	"rationalizing",	"reassessment",	"reassessments",	"reassign",	"reassigned",	"reassigning",	"reassignment",	"reassignments",	"reassigns",	"recall",	"recalled",	"recalling",	"recalls",	"recession",	"recessionary",	"recessions",	"reckless",	"recklessly",	"recklessness",	"redact",	"redacted",	"redacting",	"redaction",	"redactions",	"redefault",	"redefaulted",	"redefaults",	"redress",	"redressed",	"redresses",	"redressing",	"refusal",	"refusals",	"refuse",	"refused",	"refuses",	"refusing",	"reject",	"rejected",	"rejecting",	"rejection",	"rejections",	"rejects",	"relinquish",	"relinquished",	"relinquishes",	"relinquishing",	"relinquishment",	"relinquishments",	"reluctance",	"reluctant",	"renegotiate",	"renegotiated",	"renegotiates",	"renegotiating",	"renegotiation",	"renegotiations",	"renounce",	"renounced",	"renouncement",	"renouncements",	"renounces",	"renouncing",	"reparation",	"reparations",	"repossessed",	"repossesses",	"repossessing",	"repossession",	"repossessions",	"repudiate",	"repudiated",	"repudiates",	"repudiating",	"repudiation",	"repudiations",	"resign",	"resignation",	"resignations",	"resigned",	"resigning",	"resigns",	"restate",	"restated",	"restatement",	"restatements",	"restates",	"restating",	"restructure",	"restructured",	"restructures",	"restructuring",	"restructurings",	"retaliate",	"retaliated",	"retaliates",	"retaliating",	"retaliation",	"retaliations",	"retaliatory",	"retribution",	"retributions",	"revocation",	"revocations",	"revoke",	"revoked",	"revokes",	"revoking",	"ridicule",	"ridiculed",	"ridicules",	"ridiculing",	"riskier",	"riskiest",	"risky",	"sabotage",	"sacrifice",	"sacrificed",	"sacrifices",	"sacrificial",	"sacrificing",	"scandalous",	"scandals",	"scrutinize",	"scrutinized",	"scrutinizes",	"scrutinizing",	"scrutiny",	"secrecy",	"seize",	"seized",	"seizes",	"seizing",	"sentenced",	"sentencing",	"serious",	"seriously",	"seriousness",	"setback",	"setbacks",	"sever",	"severe",	"severed",	"severely",	"severities",	"severity",	"sharply",	"shocked",	"shortage",	"shortages",	"shortfall",	"shortfalls",	"shrinkage",	"shrinkages",	"shut",	"shutdown",	"shutdowns",	"shuts",	"shutting",	"slander",	"slandered",	"slanderous",	"slanders",	"slippage",	"slippages",	"slow",	"slowdown",	"slowdowns",	"slowed",	"slower",	"slowest",	"slowing",	"slowly",	"slowness",	"sluggish",	"sluggishly",	"sluggishness",	"solvencies",	"solvency",	"spam",	"spammers",	"spamming",	"staggering",	"stagnant",	"stagnate",	"stagnated",	"stagnates",	"stagnating",	"stagnation",	"standstill",	"standstills",	"stolen",	"stoppage",	"stoppages",	"stopped",	"stopping",	"stops",	"strain",	"strained",	"straining",	"strains",	"stress",	"stressed",	"stresses",	"stressful",	"stressing",	"stringent",	"subjected",	"subjecting",	"subjection",	"subpoena",	"subpoenaed",	"subpoenas",	"substandard",	"sue",	"sued",	"sues",	"suffer",	"suffered",	"suffering",	"suffers",	"suing",	"summoned",	"summoning",	"summons",	"summonses",	"susceptibility",	"susceptible",	"suspect",	"suspected",	"suspects",	"suspend",	"suspended",	"suspending",	"suspends",	"suspension",	"suspensions",	"suspicion",	"suspicions",	"suspicious",	"suspiciously",	"taint",	"tainted",	"tainting",	"taints",	"tampered",	"tense",	"terminate",	"terminated",	"terminates",	"terminating",	"termination",	"terminations",	"testify",	"testifying",	"threat",	"threaten",	"threatened",	"threatening",	"threatens",	"threats",	"tightening",	"tolerate",	"tolerated",	"tolerates",	"tolerating",	"toleration",	"tortuous",	"tortuously",	"tragedies",	"tragedy",	"tragic",	"tragically",	"traumatic",	"trouble",	"troubled",	"troubles",	"turbulence",	"turmoil",	"unable",	"unacceptable",	"unacceptably",	"unaccounted",	"unannounced",	"unanticipated",	"unapproved",	"unattractive",	"unauthorized",	"unavailability",	"unavailable",	"unavoidable",	"unavoidably",	"unaware",	"uncollectable",	"uncollected",	"uncollectibility",	"uncollectible",	"uncollectibles",	"uncompetitive",	"uncompleted",	"unconscionable",	"unconscionably",	"uncontrollable",	"uncontrollably",	"uncontrolled",	"uncorrected",	"uncover",	"uncovered",	"uncovering",	"uncovers",	"undeliverable",	"undelivered",	"undercapitalized",	"undercut",	"undercuts",	"undercutting",	"underestimate",	"underestimated",	"underestimates",	"underestimating",	"underestimation",	"underfunded",	"underinsured",	"undermine",	"undermined",	"undermines",	"undermining",	"underpaid",	"underpayment",	"underpayments",	"underpays",	"underperform",	"underperformance",	"underperformed",	"underperforming",	"underperforms",	"underproduced",	"underproduction",	"underreporting",	"understate",	"understated",	"understatement",	"understatements",	"understates",	"understating",	"underutilization",	"underutilized",	"undesirable",	"undesired",	"undetected",	"undetermined",	"undisclosed",	"undocumented",	"undue",	"unduly",	"uneconomic",	"uneconomical",	"uneconomically",	"unemployed",	"unemployment",	"unethical",	"unethically",	"unexcused",	"unexpected",	"unexpectedly",	"unfair",	"unfairly",	"unfavorability",	"unfavorable",	"unfavorably",	"unfavourable",	"unfeasible",	"unfit",	"unfitness",	"unforeseeable",	"unforeseen",	"unforseen",	"unfortunate",	"unfortunately",	"unfounded",	"unfriendly",	"unfulfilled",	"unfunded",	"uninsured",	"unintended",	"unintentional",	"unintentionally",	"unjust",	"unjustifiable",	"unjustifiably",	"unjustified",	"unjustly",	"unknowing",	"unknowingly",	"unlawful",	"unlawfully",	"unlicensed",	"unliquidated",	"unmarketable",	"unmerchantable",	"unmeritorious",	"unnecessarily",	"unnecessary",	"unneeded",	"unobtainable",	"unoccupied",	"unpaid",	"unperformed",	"unplanned",	"unpopular",	"unpredictability",	"unpredictable",	"unpredictably",	"unpredicted",	"unproductive",	"unprofitability",	"unprofitable",	"unqualified",	"unrealistic",	"unreasonable",	"unreasonableness",	"unreasonably",	"unreceptive",	"unrecoverable",	"unrecovered",	"unreimbursed",	"unreliable",	"unremedied",	"unreported",	"unresolved",	"unrest",	"unsafe",	"unsalable",	"unsaleable",	"unsatisfactory",	"unsatisfied",	"unsavory",	"unscheduled",	"unsellable",	"unsold",	"unsound",	"unstabilized",	"unstable",	"unsubstantiated",	"unsuccessful",	"unsuccessfully",	"unsuitability",	"unsuitable",	"unsuitably",	"unsuited",	"unsure",	"unsuspected",	"unsuspecting",	"unsustainable",	"untenable",	"untimely",	"untrusted",	"untruth",	"untruthful",	"untruthfully",	"untruthfulness",	"untruths",	"unusable",	"unwanted",	"unwarranted",	"unwelcome",	"unwilling",	"unwillingness",	"upset",	"urgency",	"urgent",	"usurious",	"usurp",	"usurped",	"usurping",	"usurps",	"usury",	"vandalism",	"verdict",	"verdicts",	"vetoed",	"victims",	"violate",	"violated",	"violates",	"violating",	"violation",	"violations",	"violative",	"violator",	"violators",	"violence",	"violent",	"violently",	"vitiate",	"vitiated",	"vitiates",	"vitiating",	"vitiation",	"voided",	"voiding",	"volatile",	"volatility",	"vulnerabilities",	"vulnerability",	"vulnerable",	"vulnerably",	"warn",	"warned",	"warning",	"warnings",	"warns",	"wasted",	"wasteful",	"wasting",	"weak",	"weaken",	"weakened",	"weakening",	"weakens",	"weaker",	"weakest",	"weakly",	"weakness",	"weaknesses",	"willfully",	"worries",	"worry",	"worrying",	"worse",	"worsen",	"worsened",	"worsening",	"worsens",	"worst",	"worthless",	"writedown",	"writedowns",	"writeoff",	"writeoffs",	"wrong",	"wrongdoing",	"wrongdoings",	"wrongful",	"wrongfully",	"wrongly"]


# define which bodies of text are relevant to our ESG study based on our personal lexicons
positive_sustainable_pattern = '|'.join(positive_sustainable_dictionary)
negative_sustainable_pattern = '|'.join(negative_sustainable_dictionary)

df["positive esg relevant"] = df["cleaned body"].str.contains(positive_sustainable_pattern)
df["negative esg relevant"] = df["cleaned body"].str.contains(negative_sustainable_pattern)

df["positive esg relevant"] = df["positive esg relevant"].map({True: 1, False: 0})
df["negative esg relevant"] = df["negative esg relevant"].map({True: -1, False: 0})

df["esg relevant"] = df["positive esg relevant"] + df["negative esg relevant"]

df.drop(columns=["positive esg relevant", "negative esg relevant"], inplace= True)


# construct polarity of sentiments based on combined lexicons
positive_loughranmcdonald_pattern = '|'.join(positive_louhgranmcdonald_dictionary)
negative_loughranmcdonald_pattern = '|'.join(negative_louhgranmcdonald_dictionary)

df["lexicon based positives"] = df["cleaned body"].str.count(positive_loughranmcdonald_pattern) + df["cleaned body"].str.count(positive_sustainable_pattern)
df["lexicon based negatives"] = df["cleaned body"].str.count(negative_loughranmcdonald_pattern) + df["cleaned body"].str.count(negative_sustainable_pattern)

df["body word count"] = df["cleaned body"].str.split().str.len()

df["own polarity"] = (df["lexicon based positives"] - df["lexicon based negatives"]) / (df["lexicon based positives"] + df["lexicon based negatives"])
df["own subjectivity"] = (df["lexicon based positives"] + df["lexicon based negatives"]) / df["body word count"]


# TextBlob Subjectivity & Polarity (alternative option for polarity)
"""
def getSubjectivity(text):
    return TextBlob(text).sentiment.subjectivity

def getPolarity(text):
    return TextBlob(text).sentiment.polarity

df["Textblob Subjectivity"] =  df["cleaned body"].apply(getSubjectivity)
df["Textblob Polarity"] = df["cleaned body"].apply(getPolarity)

df["ESG relevant Polarity"] = df[["esg relevant"]].multiply(df["own polarity"], axis="index")

print(df["esg relevant"].value_counts())
"""


#convert scraped timestamp to datetime
df["timestamp"] = df["timestamp"].astype(float)

def getTimestamp(timestamp):
    return pd.Timestamp(timestamp, unit="s")

df["datetime"] = df["timestamp"].apply(getTimestamp)

df.dropna(axis=0, inplace=True)

df.drop(columns=["timestamp", "esg relevant", "lexicon based positives", "lexicon based negatives"], inplace=True)

print(df.info())

filt_pos = df["own polarity"] > 0
filt_neg = df["own polarity"] < 0
filt_zero = df["own polarity"] == 0

df["polarity_sign"] = "undefined"
df["polarity_direction"] = "undefined"

df.loc[filt_pos, "polarity_sign"]=1
df.loc[filt_neg, "polarity_sign"]=-1
df.loc[filt_zero, "polarity_sign"]=0

df.loc[filt_pos, "polarity_direction"]="esg-bullish"
df.loc[filt_neg, "polarity_direction"]="esg-bearish"
df.loc[filt_zero, "polarity_direction"]="esg-neutral"

df = df[df["own subjectivity"] != 0]

df.to_csv("4. processed-reddit-text-thesis.csv")
